+++ src/components/common/date_picker.tsx
@@ -60,8 +60,8 @@ const PopulateDates: React.FC<IPopulateDatesProps> = ({
   ));
 
   const dateGrid = daysInMonth.map((el, index) => {
-    // eslint-disable-next-line react/jsx-key
-    if (!el) return <div className={`null-${index}`}></div>;
+     
+    if (!el) return <div key={`null-${index}`}></div>;
 
     // Info: (20250825 - Julian) 取得該日期的 timestamp (秒)
     const timestamp = new Date(`${selectedYear}/${selectedMonth}/${el.date}`).getTime() / 1000;
